refactor: rename Athlete JPA repository implementation (#36)
