+++ src/test/java/com/axel/trainingmetricsapi/service/AthleteServiceTest.java
@@ -1,7 +1,7 @@
 package com.axel.trainingmetricsapi.service;
 
 import com.axel.trainingmetricsapi.domain.Athlete;
-import com.axel.trainingmetricsapi.repository.AthleteRepository;
+import com.axel.trainingmetricsapi.domain.AthleteRepository;
 import org.instancio.Instancio;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
